requirements.txt
openai==1.35.0
langchain==0.2.11
langchain-openai==0.1.19
sqlalchemy==2.0.23
pandas==2.1.3
streamlit==1.28.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
spacy==3.7.2
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
plotly==5.17.0

# Новые зависимости для улучшений
python-json-logger==2.0.7
sqlparse==0.4.4
pydantic[email]==2.5.0
